Add prototype system devicetree for nRF5340
That's in nrf5340_system.dts, and is based on various related files in
the zephyr tree. I sometimes left out some nodes where the upstream
zephyr DTS doesn't have them, but those are annotated with 'skipped:'
comments. I also removed a bunch of node labels that would cause
duplicates in a single file.

If you run run.sh, you'll get two files: out-cpuapp.dts and
out-cpunet.dts. These are the results of running lopper on
nrf5340_system.dts to move the relevant cpu cluster to /cpus.

This is just meant as a starting point and sanity check that this SDT
is acceptable to the tool. You can't get zephyr to boot any target on
this SoC with the results.

There are various hardcoded assumptions baked into this that encode a
secure-only view of the world and should be followed up on. See
'FIXME' comments.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>