nrf5340_system: re-work peripheral memory map
This follows discussion recorded here:

https://gist.github.com/galak/51f052e7d3e2c1848d0fe7d89329a282

Reworking the nodes themselves is easy, but introduces a separate
problem of how to address the bus node for user-selectable
peripherals in a "cpus,cluster" node's address-map property.

Here we are just using the same bus twice in the same address map,
with different base addresses, creating different "views" of the
target app_us node given a base address and a size.

This produces nonsense results from run.sh, but that doesn't actually
break anything. Next steps are to put some lops and/or assists together
to transform the results to something useful.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>